+++ packages/paraglide/paraglide-js/src/compiler/compiler-options.ts
@@ -13,6 +13,7 @@ export const defaultCompilerOptions = {
 	strategy: ["cookie", "globalVariable", "baseLocale"],
 	cookieName: "PARAGLIDE_LOCALE",
 	cookieMaxAge: 60 * 60 * 24 * 400,
+	cookieDomain: "",
 } as const satisfies Partial<CompilerOptions>;
 
 export type CompilerOptions = {
@@ -103,6 +104,14 @@ export type CompilerOptions = {
 	 * @default 60 * 60 * 24 * 400
 	 */
 	cookieMaxAge?: number;
+	/**
+	 * The host to which the cookie will be sent.
+	 * If null, this defaults to the host portion of the current document location and the cookie is not available on subdomains.
+	 * Otherwise, subdomains are always included.
+	 *
+	 * @default window.location.hostname
+	 */
+	cookieDomain?: string;
 	/**
 	 * The `additionalFiles` option is an array of paths to additional files that should be copied to the output directory.
 	 *